upd(Instagram): Code correction for `Enable developer menu` patch
