+++ src/main/kotlin/app/revanced/patches/instagram/patches/developerMenu/fingerprints/ShouldAddPrefTTLFingerprint.kt
@@ -1,0 +1,9 @@
+package app.revanced.patches.instagram.patches.developerMenu.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+
+internal object ShouldAddPrefTTLFingerprint : MethodFingerprint(
+    customFingerprint = { methodDef, _ ->
+        methodDef.name == "shouldAddPrefTTL" && methodDef.definingClass == "Lcom/instagram/debug/whoptions/WhitehatOptionsFragment;"
+    },
+)
